requirements.txt
django ==5.0.3
django-environ == 0.11.2